New domain: https-everywhere
I tried my luck to close https://github.com/lgarron/badssl.com/issues/80

Please test it and also note that
https://github.com/EFForg/https-everywhere/pull/3111 has to be merged
first and a HTTPS-E version including the new ruleset has to be
published to make this work.
After this has been done we should replace  the `[unknown]` placeholder
with the version of HTTPS-E released.